requirements.txt
pymongo==3.13.0
pytz==2022.6
sqlparse==0.2.4
flask==2.2.3
django==3.2
djangorestframework==3.14.0
djongo==1.3.6 